refactor alert utils, pass settings down to createAlert fns
